requirements.txt
streamlit==1.31.1
opencv-python==4.9.0.80
mediapipe==0.10.9
yt-dlp==2025.3.25
numpy==1.26.4
pandas==2.2.1
openai==1.12.0
python-dotenv==1.0.1
pillow==10.2.0 
